serious bug fixed in annot.h
it reserved space for no_sites BEFORE calculating par_noEA
also changed type of "type" from int to long long int
new randseq still crashes at filestreams
	new file:   IN/mapping3.txt
	modified:   src/annot.cpp
	modified:   src/include/annot.h
	modified:   src/include/rnarep.h
	modified:   src/randseq.cpp
	modified:   src/rnarep.cpp
	modified:   src/test.cpp

+++ src/include/rnarep.h
@@ -133,7 +133,7 @@ namespace rnarep {
 
 			int get_no_acts();
 
-			int get_type();
+			long long int get_type();
 
 			int get_length();
 
@@ -154,7 +154,7 @@ namespace rnarep {
 			std::string seq; //sequence of replicator
 			double R; // replication rate
 			double *a; //enzymatic activities
-			unsigned int type; //an integer indicating the enzimatic activities that can be found in a replicator
+			long long int type; //an integer indicating the enzimatic activities that can be found in a replicator
 			//std::vector<int> ins;
 			//std::vector<int> subs;
 			//std::vector<int> dels;
